get ja post toimii, delete kesken

+++ backend/controllers/commentController.js
@@ -4,7 +4,6 @@ const jwt = require('../auth/auth.js');
 const getComments = async (req, res) => {
   jwt.auth(req, res, async () => {
     const groupId = req.query.id_groups;
-    console.log(req.cookies.uJwt);
     try {
       const getComments = await comment.getComments(groupId);
       console.log(getComments);
@@ -16,25 +15,42 @@ const getComments = async (req, res) => {
 };
 
 const postComments = async (req, res) => {
-  const idGroups = req.body.id_groups;
-  const idUsers = req.body.id_users;
-  const userComments = req.body.user_comments;
-  try {
-    const postComments = await comment.postComments(userComments, idUsers, idGroups);
-    res.status(200).json({ message: 'Comment posted', postComments });
-  } catch (error) {
-    res.status(404).json({ error: error.message });
-  }
+  jwt.auth(req, res, async () => {
+    const idGroups = req.body.id_groups;
+    const idUsers = req.body.id_users;
+    const userComments = req.body.user_comments;
+    try {
+      const postComments = await comment.postComments(
+        userComments,
+        idUsers,
+        idGroups
+      );
+      res.status(200).json({ message: 'Comment posted', postComments });
+    } catch (error) {
+      res.status(404).json({ error: error.message });
+    }
+  });
 };
 
 const deleteComment = async (req, res) => {
-  const commentId = req.params.id;
-  try {
-    const deletedComment = await comment.deleteComment(commentId);
-    res.status(200).json({ message: 'Comment deleted', deletedComment });
-  } catch (error) {
-    res.status(404).json({ error: error.message });
-  }
+  jwt.auth(req, res, async () => {
+    const commentId = req.params.id;
+    const userId = res.locals.userId;
+    try {
+      const commentToDelete = await comment.getCommentsByUserId(userId);
+      console.log(userId);
+      console.log(res.locals.userId);
+      const isAuthorizedToDelete = commentToDelete.some(comment => comment.id_users === userId);
+      if (isAuthorizedToDelete) {
+        const deletedComment = await comment.deleteComment(commentId);
+        res.status(200).json({ message: 'Comment deleted', deletedComment });
+      } else {
+        res.status(404).json({ error: 'Unauthorized to delete this comment' });
+      }
+    } catch (error) {
+      res.status(404).json({ error: error.message });
+    }
+  });
 };
 
 module.exports = { getComments, postComments, deleteComment };
